Renaming css modules (#85)
* Renaming CSS modules

* Still renaming CSS modules...

* Renaming CSS modules

* Using new grunt-dojo2
