+++ package.json
@@ -37,7 +37,7 @@
     "@types/jsdom": "2.0.*",
     "@types/sinon": "^1.16.31",
     "grunt": "^1.0.1",
-    "grunt-dojo2": ">=2.0.0-beta.36",
+    "grunt-dojo2": ">=2.0.0-beta.39",
     "intern": "^3.4.1",
     "jsdom": "^9.5.0",
     "sinon": "^1.17.6",
