+++ src/button/Button.ts
@@ -2,7 +2,7 @@ import { WidgetBase } from '@dojo/widget-core/WidgetBase';
 import { DNode } from '@dojo/widget-core/interfaces';
 import { ThemeableMixin, ThemeableProperties, theme } from '@dojo/widget-core/mixins/Themeable';
 import { v } from '@dojo/widget-core/d';
-import * as css from './styles/button.css';
+import * as css from './styles/button.m.css';
 
 export type ButtonType = 'submit' | 'reset' | 'button' | 'menu';
