+++ src/combobox/ResultMenu.ts
@@ -5,7 +5,7 @@ import { v, w } from '@dojo/widget-core/d';
 import { DNode, WNode } from '@dojo/widget-core/interfaces';
 import { ResultItemProperties } from './ResultItem';
 
-import * as css from './styles/comboBox.css';
+import * as css from './styles/comboBox.m.css';
 
 /**
  * @type ResultMenuProperties
