+++ src/combobox/styles/comboBox.m.css.d.ts
@@ -1,10 +1,10 @@
 export const root: string;
+export const controls: string;
 export const input: string;
-export const arrow: string;
+export const clearable: string;
 export const results: string;
 export const result: string;
+export const disabledResult: string;
 export const selectedResult: string;
 export const clear: string;
-export const clearable: string;
-export const disabledResult: string;
-export const controls: string;
+export const arrow: string;
