+++ src/combobox/tests/unit/ResultItem.ts
@@ -2,7 +2,7 @@ import * as registerSuite from 'intern!object';
 import * as assert from 'intern/chai!assert';
 import { VNode } from '@dojo/interfaces/vdom';
 import ResultItem from '../../ResultItem';
-import * as css from '../../styles/comboBox.css';
+import * as css from '../../styles/comboBox.m.css';
 import { assign } from '@dojo/core/lang';
 
 function props(props = {}) {
