+++ src/common/styles/widgets.css
@@ -1,11 +1,11 @@
-@import 'animations.css';
-@import 'base.css';
-@import '../../dialog/styles/dialog.css';
-@import '../../slidepane/styles/slidePane.css';
-@import '../../button/styles/button.css';
-@import '../../textinput/styles/textinput.css';
-@import '../../textarea/styles/textarea.css';
-@import '../../checkbox/styles/checkbox.css';
-@import '../../radio/styles/radio.css';
-@import '../../slider/styles/slider.css';
-@import '../../combobox/styles/comboBox.css';
+@import 'animations.m.css';
+@import 'base.m.css';
+@import '../../dialog/styles/dialog.m.css';
+@import '../../slidepane/styles/slidePane.m.css';
+@import '../../button/styles/button.m.css';
+@import '../../textinput/styles/textinput.m.css';
+@import '../../textarea/styles/textarea.m.css';
+@import '../../checkbox/styles/checkbox.m.css';
+@import '../../radio/styles/radio.m.css';
+@import '../../slider/styles/slider.m.css';
+@import '../../combobox/styles/comboBox.m.css';
