+++ src/dialog/styles/dialog.m.css.d.ts
