+++ src/label/Label.ts
@@ -3,7 +3,7 @@ import { DNode } from '@dojo/widget-core/interfaces';
 import { ThemeableMixin, ThemeableProperties, theme, ClassesFunctionChain } from '@dojo/widget-core/mixins/Themeable';
 import { v } from '@dojo/widget-core/d';
 import { assign } from '@dojo/core/lang';
-import * as baseCss from '../common/styles/base.css';
+import * as baseCss from '../common/styles/base.m.css';
 
 /**
  * Label settings for form label text content, position (before or after), and visibility
