+++ src/slidepane/SlidePane.ts
@@ -3,8 +3,8 @@ import { WidgetBase } from '@dojo/widget-core/WidgetBase';
 import { ThemeableMixin, ThemeableProperties, theme } from '@dojo/widget-core/mixins/Themeable';
 import { v } from '@dojo/widget-core/d';
 
-import * as css from './styles/slidePane.css';
-import * as animations from '../common/styles/animations.css';
+import * as css from './styles/slidePane.m.css';
+import * as animations from '../common/styles/animations.m.css';
 
 /**
  * Enum for left / right alignment
