+++ src/slidepane/styles/slidePane.m.css
