+++ src/slidepane/styles/slidePane.m.css.d.ts
