+++ src/slider/styles/slider.m.css
