+++ src/slider/styles/slider.m.css.d.ts
