+++ src/textarea/Textarea.ts
@@ -3,7 +3,7 @@ import { DNode } from '@dojo/widget-core/interfaces';
 import { ThemeableMixin, ThemeableProperties, theme } from '@dojo/widget-core/mixins/Themeable';
 import { v, w } from '@dojo/widget-core/d';
 import Label, { LabelOptions } from '../label/Label';
-import * as css from './styles/textarea.css';
+import * as css from './styles/textarea.m.css';
 
 /**
  * @type TextareaProperties
